Recreating in Flask Classy
