+++ .gitignore
@@ -2,3 +2,6 @@
 .idea
 
 config.py
+*.pyc
+
+*.pyc
